consolidate test utilities
